Solution to Calculate Area and Perimeter of Rectangles

+++ Calculate Area and Perimeter of Rectangles@@67877d61dc107486ad5c274b/code.c
@@ -15,7 +15,7 @@ int main(){
         scanf("%f %f", &r[i].length, &r[i].breadth);
     }
     for(int i =0;i<a;i++){
-        int area = r[i].length * r[i].breadth;
+        float area = r[i].length * r[i].breadth;
         float perimeter = r[i].length + r[i].breadth;
         printf("Rectangle %d: Area = %.2f, Perimeter = %.2f\n", i+1, area, perimeter);
     }
